+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Mean.scala
@@ -20,7 +20,7 @@ import java.nio.ByteOrder
 import com.intel.analytics.bigdl.Module
 import com.intel.analytics.bigdl.nn.abstractnn.{AbstractModule, Activity}
 import com.intel.analytics.bigdl.nn.Sequential
-import com.intel.analytics.bigdl.nn.ops.Mean
+import com.intel.analytics.bigdl.nn.tf.Mean
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
