+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Relu6.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.ReLU6
+import com.intel.analytics.bigdl.nn.tf.ReLU6
 import com.intel.analytics.bigdl.nn.tf.Log1p
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
