+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Softplus.scala
@@ -18,8 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.SoftPlus
-import com.intel.analytics.bigdl.nn.ops.Relu6Grad
+import com.intel.analytics.bigdl.nn.tf.SoftPlus
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
 import com.intel.analytics.bigdl.utils.tf.loaders.Utils.getType
